+++ services/course-material/src/components/ContentRenderer/core/common/GlossaryTooltip/index.tsx
