+++ services/course-material/src/components/ContentRenderer/core/common/Paragraph/index.tsx
@@ -1,75 +1,19 @@
 import { css } from "@emotion/css"
 import { diffChars } from "diff"
-import KaTex from "katex"
 import dynamic from "next/dynamic"
-import React, { useEffect, useState } from "react"
-import * as ReactDOMServer from "react-dom/server"
+import React, { useContext, useEffect, useState } from "react"
 import { useMemo } from "use-memo-one"
 
 import { BlockRendererProps } from "../../.."
 import { ParagraphAttributes } from "../../../../../../types/GutenbergBlockAttributes"
-import { Term } from "../../../../../shared-module/bindings"
+import { GlossaryContext } from "../../../../../contexts/GlossaryContext"
 import DiffFormatter from "../../../../../shared-module/components/DiffFormatter"
 import colorMapper from "../../../../../styles/colorMapper"
 import fontSizeMapper from "../../../../../styles/fontSizeMapper"
-import { sanitizeCourseMaterialHtml } from "../../../../../utils/sanitizeCourseMaterialHtml"
-
-import Tooltip from "./GlossaryTooltip"
+import { parseText } from "../../../util/textParsing"
 
 const Paragraph = dynamic(() => import("./BasicParagraph"))
 const LatexParagraph = dynamic(() => import("./LatexParagraph"))
-
-const LATEX_REGEX = /\[latex\](.*?)\[\/latex\]/g
-const LATEX_CITE_REGEX = /\\cite{(.*?)}/g
-const HTML_ESCAPED_AMPERSAND = "&amp;"
-const KATEX_OUTPUT_FORMAT = "htmlAndMathml"
-const IGNORE_CASE = "i"
-/**
- *
- * @param data HTML-content from the server
- * @returns HTML as string in which "[latex] ... [/latex]" will be replaced with katex
- */
-const convertToLatex = (data: string) => {
-  let count = 0
-  const converted = data.replace(LATEX_REGEX, (_, latex) => {
-    // Convert ampersand back to special symbol. This is needed e.g. in matrices
-    const processed = latex.replaceAll(HTML_ESCAPED_AMPERSAND, "&")
-    count++
-    return KaTex.renderToString(processed, {
-      throwOnError: false,
-      output: KATEX_OUTPUT_FORMAT,
-    })
-  })
-
-  return { count, converted }
-}
-
-const generateToolTip = (term: Term) => {
-  return ReactDOMServer.renderToString(<Tooltip term={term} />)
-}
-
-const parseGlossary = (data: string, glossary: Term[]): string => {
-  let parsed = data
-
-  glossary.forEach((item) => {
-    // eslint-disable-next-line i18next/no-literal-string
-    const regexString = `[^a-zA-Z0-9](${item.term})[^a-zA-Z0-9]`
-    parsed = parsed.replace(new RegExp(regexString, IGNORE_CASE), (content, _) =>
-      generateToolTip({ ...item, term: content }),
-    )
-  })
-
-  return parsed
-}
-
-const parseCitation = (data: string) => {
-  const converted = data.replace(
-    LATEX_CITE_REGEX,
-    // eslint-disable-next-line i18next/no-literal-string
-    (_, citationId) => `<sup class="reference" data-citation-id=${citationId}>[?]</sup>`,
-  )
-  return converted
-}
 
 const hasDropCap = css`
   :first-letter {
@@ -83,18 +27,9 @@ const hasDropCap = css`
   }
 `
 
-const parseText = (content: string, terms: Term[]) => {
-  const sanitizedHTML = sanitizeCourseMaterialHtml(content)
-  const { count, converted: parsedLatex } = convertToLatex(sanitizedHTML)
-  const parsedCitation = parseCitation(parsedLatex)
-  const parsedGlossary = parseGlossary(parsedCitation, terms ?? [])
-
-  return { count, parsedText: parsedGlossary }
-}
-
 const ParagraphBlock: React.FC<
   React.PropsWithChildren<BlockRendererProps<ParagraphAttributes>>
-> = ({ data, id, editing, selectedBlockId, setEdits, terms }) => {
+> = ({ data, id, editing, selectedBlockId, setEdits }) => {
   const {
     textColor,
     backgroundColor,
@@ -113,6 +48,7 @@ const ParagraphBlock: React.FC<
   // eslint-disable-next-line i18next/no-literal-string
   const bgColor = colorMapper(backgroundColor, "unset")
   const [editedContent, setEditedContent] = useState(data.attributes.content)
+  const { terms } = useContext(GlossaryContext)
 
   // edited content should not persist between edit proposals
   // reset edited content when no longer editing
@@ -186,8 +122,7 @@ const ParagraphBlock: React.FC<
       )
     }
   }
-
-  const { count, parsedText } = parseText(content, terms ?? [])
+  const { count, parsedText } = parseText(content, terms)
   const P = count > 0 ? LatexParagraph : Paragraph
 
   return (
