+++ services/course-material/src/components/Page.tsx
@@ -1,6 +1,7 @@
 import { useQuery } from "@tanstack/react-query"
 import React, { useContext, useState } from "react"
 
+import { GlossaryContext, GlossaryState } from "../contexts/GlossaryContext"
 import PageContext from "../contexts/PageContext"
 import useSelectedBlockId from "../hooks/useSelectedBlockId"
 import { Block, fetchGlossary } from "../services/backend"
@@ -45,51 +46,54 @@ const Page: React.FC<React.PropsWithChildren<Props>> = ({ onRefresh, organizatio
     return <ErrorBanner variant={"readOnly"} error={glossary.error} />
   }
 
+  const glossaryState: GlossaryState = { terms: glossary.data }
+
   return (
-    <>
-      {pageContext.settings &&
-        pageContext.settings.current_course_instance_id !== pageContext.instance?.id && (
-          <UserOnWrongCourseNotification
-            correctCourseId={pageContext.settings?.current_course_id}
-            organizationSlug={organizationSlug}
+    <GlossaryContext.Provider value={glossaryState}>
+      <>
+        {pageContext.settings &&
+          pageContext.settings.current_course_instance_id !== pageContext.instance?.id && (
+            <UserOnWrongCourseNotification
+              correctCourseId={pageContext.settings?.current_course_id}
+              organizationSlug={organizationSlug}
+            />
+          )}
+        {courseId && <SelectCourseInstanceModal onClose={onRefresh} />}
+        {courseId && pageId && (
+          <FeedbackHandler
+            courseId={courseId}
+            pageId={pageId}
+            onEnterEditProposalMode={() => {
+              setEditingMaterial(true)
+            }}
+            onExitEditProposalMode={() => {
+              setEditingMaterial(false)
+              setEdits(new Map())
+            }}
+            selectedBlockId={selectedBlockId}
+            clearSelectedBlockId={clearSelectedBlockId}
+            edits={edits}
           />
         )}
-      {courseId && <SelectCourseInstanceModal onClose={onRefresh} />}
-      {courseId && pageId && (
-        <FeedbackHandler
-          courseId={courseId}
-          pageId={pageId}
-          onEnterEditProposalMode={() => {
-            setEditingMaterial(true)
-          }}
-          onExitEditProposalMode={() => {
-            setEditingMaterial(false)
-            setEdits(new Map())
-          }}
-          selectedBlockId={selectedBlockId}
-          clearSelectedBlockId={clearSelectedBlockId}
-          edits={edits}
-        />
-      )}
-      {pageContext.pageData?.content && Boolean(pageContext.pageData?.chapter_id) && (
-        <HeadingsNavigation />
-      )}
-      {/* TODO: Better type for Page.content in bindings. */}
-      <div id="content" className={inlineColorStyles}>
-        <ContentRenderer
-          glossary={glossary.data}
-          data={(pageContext.pageData?.content as Array<Block<unknown>>) ?? []}
-          editing={editingMaterial}
-          selectedBlockId={selectedBlockId}
-          setEdits={setEdits}
-          isExam={pageContext.exam !== null}
-        />
-      </div>
-      {pageContext.pageData?.chapter_id && <NavigationContainer />}
-      {pageContext.pageData?.course_id && (
-        <ReferenceList courseId={pageContext.pageData.course_id} />
-      )}
-    </>
+        {pageContext.pageData?.content && Boolean(pageContext.pageData?.chapter_id) && (
+          <HeadingsNavigation />
+        )}
+        {/* TODO: Better type for Page.content in bindings. */}
+        <div id="content" className={inlineColorStyles}>
+          <ContentRenderer
+            data={(pageContext.pageData?.content as Array<Block<unknown>>) ?? []}
+            editing={editingMaterial}
+            selectedBlockId={selectedBlockId}
+            setEdits={setEdits}
+            isExam={pageContext.exam !== null}
+          />
+        </div>
+        {pageContext.pageData?.chapter_id && <NavigationContainer />}
+        {pageContext.pageData?.course_id && (
+          <ReferenceList courseId={pageContext.pageData.course_id} />
+        )}
+      </>
+    </GlossaryContext.Provider>
   )
 }
